+++ src/salespyforce/errors/exceptions.py
@@ -4,13 +4,19 @@
 :Synopsis:          Collection of exception classes relating to the SalesPyForce library
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     25 Feb 2026
+:Modified Date:     27 Feb 2026
 """
 
-#################
-# Base Exception
-#################
+from __future__ import annotations
 
+from typing import Optional, Union
+
+from ..constants import _EXCEPTION_CLASSES, API_REQUEST_TYPES
+
+
+# -----------------------------
+# Base Exception
+# -----------------------------
 
 # Define base exception class
 class SalesPyForceError(Exception):
@@ -18,20 +24,18 @@ class SalesPyForceError(Exception):
     pass
 
 
-#####################
+# -----------------------------
 # General Exceptions
-#####################
-
+# -----------------------------
 
 class CurrentlyUnsupportedError(SalesPyForceError):
     """This exception is used when a feature or functionality being used is currently unsupported."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "This feature is currently unsupported at this time."
+        default_msg = 'This feature is currently unsupported at this time.'
         if not (args or kwargs):
             args = (default_msg,)
-        elif 'message' in kwargs:
-            args =(kwargs['message'],)
+        elif _EXCEPTION_CLASSES._MESSAGE in kwargs:
+            args =(kwargs[_EXCEPTION_CLASSES._MESSAGE],)
         else:
             custom_msg = f"The '{args[0]}' {default_msg.split('This ')[1]}"
             args = (custom_msg,)
@@ -41,18 +45,20 @@ class CurrentlyUnsupportedError(SalesPyForceError):
 class DataMismatchError(SalesPyForceError):
     """This exception is used when there is a mismatch between two data sources."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "A data mismatch was found with the data sources."
+        default_msg = 'A data mismatch was found with the data sources.'
         if not (args or kwargs):
             args = (default_msg,)
-        elif 'data' in kwargs:
+        elif _EXCEPTION_CLASSES._DATA in kwargs:
             multi_types = [list, tuple, set]
-            if type(kwargs['data']) == str:
-                custom_msg = f"{default_msg.split('the data')[0]}the '{kwargs['data']}'{default_msg.split('with the')[1]}"
+            if isinstance(kwargs[_EXCEPTION_CLASSES._DATA], str):
+                custom_msg = (f"{default_msg.split('the data')[0]}the '{kwargs[_EXCEPTION_CLASSES._DATA]}'"
+                              f"{default_msg.split('with the')[1]}")
                 custom_msg = custom_msg.replace('sources', 'source')
                 args = (custom_msg,)
-            elif type(kwargs['data']) in multi_types and len(kwargs['data']) == 2:
-                custom_section = f"'{kwargs['data'][0]}' and '{kwargs['data'][1]}'"
+            elif (type(kwargs[_EXCEPTION_CLASSES._DATA]) in multi_types
+                  and len(kwargs[_EXCEPTION_CLASSES._DATA]) == 2):
+                custom_section = (f"'{kwargs[_EXCEPTION_CLASSES._DATA][0]}' and '"
+                                  f"{kwargs[_EXCEPTION_CLASSES._DATA][1]}'")
                 custom_msg = f"{default_msg.split('data sources')[0]}{custom_section}{default_msg.split('with the')[1]}"
                 args = (custom_msg,)
         super().__init__(*args)
@@ -61,12 +67,12 @@ class DataMismatchError(SalesPyForceError):
 class InvalidParameterError(SalesPyForceError):
     """This exception is used when an invalid parameter is provided."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "The parameter that was provided is invalid."
+        default_msg = 'The parameter that was provided is invalid.'
         if not (args or kwargs):
             args = (default_msg,)
-        elif 'val' in kwargs:
-            custom_msg = f"{default_msg.split('parameter ')[0]}'{kwargs['val']}'{default_msg.split('The')[1]}"
+        elif _EXCEPTION_CLASSES._VAL in kwargs:
+            custom_msg = (f"{default_msg.split('parameter ')[0]}'{kwargs[_EXCEPTION_CLASSES._VAL]}'"
+                          f"{default_msg.split('The')[1]}")
             args = (custom_msg,)
         super().__init__(*args)
 
@@ -74,12 +80,12 @@ class InvalidParameterError(SalesPyForceError):
 class InvalidFieldError(SalesPyForceError):
     """This exception is used when an invalid field is provided."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "The field that was provided is invalid."
+        default_msg = 'The field that was provided is invalid.'
         if not (args or kwargs):
             args = (default_msg,)
-        elif 'val' in kwargs:
-            custom_msg = f"{default_msg.split('field ')[0]}'{kwargs['val']}'{default_msg.split('The')[1]}"
+        elif _EXCEPTION_CLASSES._VAL in kwargs:
+            custom_msg = (f"{default_msg.split('field ')[0]}'{kwargs[_EXCEPTION_CLASSES._VAL]}'"
+                          f"{default_msg.split('The')[1]}")
             args = (custom_msg,)
         super().__init__(*args)
 
@@ -87,12 +93,12 @@ class InvalidFieldError(SalesPyForceError):
 class InvalidURLError(SalesPyForceError):
     """This exception is used when a provided URL is invalid."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "The provided URL is invalid"
+        default_msg = 'The provided URL is invalid'
         if not (args or kwargs):
             args = (default_msg,)
-        elif 'url' in kwargs:
-            custom_msg = f"{default_msg.split('is')[0]}'{kwargs['url']}'{default_msg.split('URL')[1]}"
+        elif _EXCEPTION_CLASSES._URL in kwargs:
+            custom_msg = (f"{default_msg.split('is')[0]}'{kwargs[_EXCEPTION_CLASSES._URL]}'"
+                          f"{default_msg.split('URL')[1]}")
             args = (custom_msg,)
         super().__init__(*args)
 
@@ -100,20 +106,20 @@ class InvalidURLError(SalesPyForceError):
 class MissingRequiredDataError(SalesPyForceError):
     """This exception is used when a function or method is missing one or more required arguments."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "Missing one or more required parameters"
-        init_msg = "The object failed to initialize as it is missing one or more required arguments."
+        default_msg = 'Missing one or more required parameters'
+        init_msg = 'The object failed to initialize as it is missing one or more required arguments.'
         param_msg = "The required parameter 'PARAMETER_NAME' is not defined"
         if not (args or kwargs):
             args = (default_msg,)
-        elif 'init' in args or 'initialize' in args:
-            if 'object' in kwargs:
-                custom_msg = f"{init_msg.split('object')[0]}'{kwargs['object']}'{init_msg.split('The')[1]}"
+        elif _EXCEPTION_CLASSES._INIT in args or _EXCEPTION_CLASSES._INITIALIZE in args:
+            if _EXCEPTION_CLASSES._OBJECT in kwargs:
+                custom_msg = (f"{init_msg.split(_EXCEPTION_CLASSES._OBJECT)[0]}'"
+                              f"{kwargs[_EXCEPTION_CLASSES._OBJECT]}'{init_msg.split('The')[1]}")
                 args = (custom_msg,)
             else:
                 args = (init_msg,)
-        elif 'param' in kwargs:
-            args = (param_msg.replace('PARAMETER_NAME', kwargs['param']),)
+        elif _EXCEPTION_CLASSES._PARAM in kwargs:
+            args = (param_msg.replace('PARAMETER_NAME', kwargs[_EXCEPTION_CLASSES._PARAM]),)
         else:
             args = (default_msg,)
         super().__init__(*args)
@@ -122,41 +128,91 @@ class MissingRequiredDataError(SalesPyForceError):
 class UnknownFileTypeError(SalesPyForceError):
     """This exception is used when a file type for a given file cannot be identified."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "The file type of the given file path cannot be identified."
+        default_msg = 'The file type of the given file path cannot be identified.'
         if not (args or kwargs):
             args = (default_msg,)
-        elif 'file' in kwargs:
-            custom_msg = f"{default_msg.split('path')[0]}'{kwargs['file']}'{default_msg.split('path')[1]}"
+        elif _EXCEPTION_CLASSES._FILE in kwargs:
+            delimiter = 'path'
+            custom_msg = (f"{default_msg.split(delimiter)[0]}'{kwargs[_EXCEPTION_CLASSES._FILE]}'"
+                          f"{default_msg.split(delimiter)[1]}")
             args = (custom_msg,)
         super().__init__(*args)
 
 
-#########################
+# -----------------------------
 # Generic API Exceptions
-#########################
-
+# -----------------------------
 
 class APIConnectionError(SalesPyForceError):
     """This exception is used when the API query could not be completed due to connection aborts and/or timeouts."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "The API query could not be completed due to connection aborts and/or timeouts."
+        default_msg = 'The API query could not be completed due to connection aborts and/or timeouts.'
         if not (args or kwargs):
             args = (default_msg,)
         super().__init__(*args)
 
 
 class APIRequestError(SalesPyForceError):
-    """This exception is used for generic API request errors when there isn't a more specific exception.
-
-    .. versionchanged:: 4.5.0
-       Fixed an issue with the default message.
-    """
+    """This exception is used for generic API request errors when there isn't a more specific exception."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "The API request did not return a successful response."
+        default_msg = _EXCEPTION_CLASSES._API_DEFAULT_MSG.format(type='API')
         if not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class GETRequestError(SalesPyForceError):
+    """This exception is used for generic GET request errors when there isn't a more specific exception."""
+    def __init__(self, *args, **kwargs):
+        default_msg = _EXCEPTION_CLASSES._API_DEFAULT_MSG.format(type=API_REQUEST_TYPES.GET)
+        if _EXCEPTION_CLASSES._STATUS_CODE in kwargs or _EXCEPTION_CLASSES._MESSAGE in kwargs:
+            custom_msg = _construct_api_custom_message(_request_type=API_REQUEST_TYPES.GET,
+                                                       _message=kwargs.get(_EXCEPTION_CLASSES._MESSAGE, None),
+                                                       _status_code=kwargs.get(_EXCEPTION_CLASSES._STATUS_CODE, None))
+            args = (custom_msg,)
+        elif not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class PATCHRequestError(SalesPyForceError):
+    """This exception is used for generic PATCH request errors when there isn't a more specific exception."""
+    def __init__(self, *args, **kwargs):
+        default_msg = _EXCEPTION_CLASSES._API_DEFAULT_MSG.format(type=API_REQUEST_TYPES.PATCH)
+        if _EXCEPTION_CLASSES._STATUS_CODE in kwargs or _EXCEPTION_CLASSES._MESSAGE in kwargs:
+            custom_msg = _construct_api_custom_message(_request_type=API_REQUEST_TYPES.PATCH,
+                                                       _message=kwargs.get(_EXCEPTION_CLASSES._MESSAGE, None),
+                                                       _status_code=kwargs.get(_EXCEPTION_CLASSES._STATUS_CODE, None))
+            args = (custom_msg,)
+        elif not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class POSTRequestError(SalesPyForceError):
+    """This exception is used for generic POST request errors when there isn't a more specific exception."""
+    def __init__(self, *args, **kwargs):
+        default_msg = _EXCEPTION_CLASSES._API_DEFAULT_MSG.format(type=API_REQUEST_TYPES.POST)
+        if _EXCEPTION_CLASSES._STATUS_CODE in kwargs or _EXCEPTION_CLASSES._MESSAGE in kwargs:
+            custom_msg = _construct_api_custom_message(_request_type=API_REQUEST_TYPES.POST,
+                                                       _message=kwargs.get(_EXCEPTION_CLASSES._MESSAGE, None),
+                                                       _status_code=kwargs.get(_EXCEPTION_CLASSES._STATUS_CODE, None))
+            args = (custom_msg,)
+        elif not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class PUTRequestError(SalesPyForceError):
+    """This exception is used for generic PUT request errors when there isn't a more specific exception."""
+    def __init__(self, *args, **kwargs):
+        default_msg = _EXCEPTION_CLASSES._API_DEFAULT_MSG.format(type=API_REQUEST_TYPES.PUT)
+        if _EXCEPTION_CLASSES._STATUS_CODE in kwargs or _EXCEPTION_CLASSES._MESSAGE in kwargs:
+            custom_msg = _construct_api_custom_message(_request_type=API_REQUEST_TYPES.PUT,
+                                                       _message=kwargs.get(_EXCEPTION_CLASSES._MESSAGE, None),
+                                                       _status_code=kwargs.get(_EXCEPTION_CLASSES._STATUS_CODE, None))
+            args = (custom_msg,)
+        elif not (args or kwargs):
             args = (default_msg,)
         super().__init__(*args)
 
@@ -164,20 +220,20 @@ class APIRequestError(SalesPyForceError):
 class DELETERequestError(SalesPyForceError):
     """This exception is used for generic DELETE request errors when there isn't a more specific exception."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "The DELETE request did not return a successful response."
-        if not (args or kwargs):
+        default_msg = _EXCEPTION_CLASSES._API_DEFAULT_MSG.format(type=API_REQUEST_TYPES.DELETE)
+        if _EXCEPTION_CLASSES._STATUS_CODE in kwargs or _EXCEPTION_CLASSES._MESSAGE in kwargs:
+            custom_msg = _construct_api_custom_message(_request_type=API_REQUEST_TYPES.DELETE,
+                                                       _message=kwargs.get(_EXCEPTION_CLASSES._MESSAGE, None),
+                                                       _status_code=kwargs.get(_EXCEPTION_CLASSES._STATUS_CODE, None))
+            args = (custom_msg,)
+        elif not (args or kwargs):
             args = (default_msg,)
         super().__init__(*args)
 
 
 class FeatureNotConfiguredError(SalesPyForceError):
-    """This exception is used when an API request fails because a feature is not configured.
-
-    .. versionadded:: 4.0.0
-    """
+    """This exception is used when an API request fails because a feature is not configured."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         exc_msg = "The feature is not configured."
         if 'identifier' in kwargs or 'feature' in kwargs:
             if 'identifier' in kwargs:
@@ -190,38 +246,9 @@ class FeatureNotConfiguredError(SalesPyForceError):
         super().__init__(*args)
 
 
-class GETRequestError(SalesPyForceError):
-    """This exception is used for generic GET request errors when there isn't a more specific exception.
-
-    .. versionchanged:: 3.2.0
-       Enabled the ability to optionally pass ``status_code`` and/or ``message`` arguments.
-    """
-    def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "The GET request did not return a successful response."
-        custom_msg = "The GET request failed with the following message:"
-        if 'status_code' in kwargs or 'message' in kwargs:
-            if 'status_code' in kwargs:
-                status_code_msg = f"returned the {kwargs['status_code']} status code"
-                custom_msg = custom_msg.replace('failed', status_code_msg)
-            if 'message' in kwargs:
-                custom_msg = f"{custom_msg} {kwargs['message']}"
-            else:
-                custom_msg = custom_msg.split(' with the following')[0] + "."
-            args = (custom_msg,)
-        elif not (args or kwargs):
-            args = (default_msg,)
-        super().__init__(*args)
-
-
 class InvalidEndpointError(SalesPyForceError):
-    """This exception is used when an invalid API endpoint / service is provided.
-
-    .. versionchanged:: 5.1.2
-       Removed part of the default message that was specifically for Khoros JX, which is obsolete.
-    """
+    """This exception is used when an invalid API endpoint / service is provided."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         default_msg = "The supplied endpoint for the API is not recognized."
         if not (args or kwargs):
             args = (default_msg,)
@@ -231,7 +258,6 @@ class InvalidEndpointError(SalesPyForceError):
 class InvalidLookupTypeError(SalesPyForceError):
     """This exception is used when an invalid API lookup type is provided."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         default_msg = "The supplied lookup type for the API is not recognized. (Examples of valid " + \
                       "lookup types include 'id' and 'email')"
         if not (args or kwargs):
@@ -240,12 +266,8 @@ class InvalidLookupTypeError(SalesPyForceError):
 
 
 class InvalidPayloadValueError(SalesPyForceError):
-    """This exception is used when an invalid value is provided for a payload field.
-
-    .. versionadded:: 2.6.0
-    """
+    """This exception is used when an invalid value is provided for a payload field."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         default_msg = "An invalid payload value was provided."
         custom_msg = "The invalid payload value 'X' was provided for the 'Y' field."
         if not (args or kwargs):
@@ -262,7 +284,6 @@ class InvalidPayloadValueError(SalesPyForceError):
 class InvalidRequestTypeError(SalesPyForceError):
     """This exception is used when an invalid API request type is provided."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         default_msg = "The supplied request type for the API is not recognized. (Examples of valid " + \
                       "request types include 'POST' and 'PUT')"
         if not (args or kwargs):
@@ -273,7 +294,6 @@ class InvalidRequestTypeError(SalesPyForceError):
 class LookupMismatchError(SalesPyForceError):
     """This exception is used when a lookup value doesn't match the supplied lookup type."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         default_msg = "The supplied lookup type for the API does not match the value that was provided."
         if not (args or kwargs):
             args = (default_msg,)
@@ -283,7 +303,6 @@ class LookupMismatchError(SalesPyForceError):
 class NotFoundResponseError(SalesPyForceError):
     """This exception is used when an API query returns a 404 response and there isn't a more specific class."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         default_msg = "The API query returned a 404 response."
         if not (args or kwargs):
             args = (default_msg,)
@@ -291,12 +310,8 @@ class NotFoundResponseError(SalesPyForceError):
 
 
 class PayloadMismatchError(SalesPyForceError):
-    """This exception is used when more than one payload is supplied for an API request.
-
-    .. versionadded:: 3.2.0
-    """
+    """This exception is used when more than one payload is supplied for an API request."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         default_msg = "More than one payload was provided for the API call when only one is permitted."
         if not (args or kwargs):
             args = (default_msg,)
@@ -306,63 +321,13 @@ class PayloadMismatchError(SalesPyForceError):
         super().__init__(*args)
 
 
-class POSTRequestError(SalesPyForceError):
-    """This exception is used for generic POST request errors when there isn't a more specific exception.
-
-    .. versionchanged:: 3.2.0
-       Enabled the ability to optionally pass ``status_code`` and/or ``message`` arguments.
-    """
-    def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "The POST request did not return a successful response."
-        custom_msg = "The POST request failed with the following message:"
-        if 'status_code' in kwargs or 'message' in kwargs:
-            if 'status_code' in kwargs:
-                status_code_msg = f"returned the {kwargs['status_code']} status code"
-                custom_msg = custom_msg.replace('failed', status_code_msg)
-            if 'message' in kwargs:
-                custom_msg = f"{custom_msg} {kwargs['message']}"
-            else:
-                custom_msg = custom_msg.split(' with the following')[0] + "."
-            args = (custom_msg,)
-        elif not (args or kwargs):
-            args = (default_msg,)
-        super().__init__(*args)
-
-
-class PUTRequestError(SalesPyForceError):
-    """This exception is used for generic PUT request errors when there isn't a more specific exception.
-
-    .. versionchanged:: 3.2.0
-       Enabled the ability to optionally pass ``status_code`` and/or ``message`` arguments.
-    """
-    def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
-        default_msg = "The PUT request did not return a successful response."
-        custom_msg = "The PUT request failed with the following message:"
-        if 'status_code' in kwargs or 'message' in kwargs:
-            if 'status_code' in kwargs:
-                status_code_msg = f"returned the {kwargs['status_code']} status code"
-                custom_msg = custom_msg.replace('failed', status_code_msg)
-            if 'message' in kwargs:
-                custom_msg = f"{custom_msg} {kwargs['message']}"
-            else:
-                custom_msg = custom_msg.split(' with the following')[0] + "."
-            args = (custom_msg,)
-        elif not (args or kwargs):
-            args = (default_msg,)
-        super().__init__(*args)
-
-
-####################
+# -----------------------------
 # Helper Exceptions
-####################
-
+# -----------------------------
 
 class InvalidHelperFileTypeError(SalesPyForceError, ValueError):
     """This exception is used when an invalid file type is provided for the helper file."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         default_msg = "The helper configuration file can only have the 'yml', 'yaml', or 'json' file type."
         if not (args or kwargs):
             args = (default_msg,)
@@ -372,7 +337,6 @@ class InvalidHelperFileTypeError(SalesPyForceError, ValueError):
 class InvalidHelperArgumentsError(SalesPyForceError):
     """This exception is used when the helper function was supplied arguments instead of keyword arguments."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         default_msg = "The helper configuration file only accepts basic keyword arguments. (e.g. arg_name='arg_value')"
         if not (args or kwargs):
             args = (default_msg,)
@@ -382,8 +346,46 @@ class InvalidHelperArgumentsError(SalesPyForceError):
 class HelperFunctionNotFoundError(SalesPyForceError):
     """This exception is used when a function referenced in the helper config file does not exist."""
     def __init__(self, *args, **kwargs):
-        """This method defines the default or custom message for the exception."""
         default_msg = "The function referenced in the helper configuration file could not be found."
         if not (args or kwargs):
             args = (default_msg,)
         super().__init__(*args)
+
+
+def _construct_api_custom_message(
+        _request_type: str,
+        _message: Optional[str] = None,
+        _status_code: Union[Optional[str], Optional[int]] = None
+) -> str:
+    """This function constructions the exception message for an API-related exception class.
+
+    .. versionadded:: 1.5.0
+
+    :param _request_type: The associated API request type (``GET``, ``PATCH``, ``POST``, ``PUT``, or ``DELETE``)
+    :type _request_type: str
+    :param _message: A specific message to append to the base message (optional)
+    :type _message: str
+    :param _status_code: The status code returned from the API request (optional)
+    :type _status_code: str, int, None
+    :returns: The constructed custom message to use when raising the exception
+    """
+    # Define the base custom message
+    _custom_msg = _EXCEPTION_CLASSES._API_CUSTOM_MSG.format(type=_request_type.upper())
+
+    # Define the status code custom message if a status code was provided
+    if _status_code:
+        _status_code_msg = f'returned the {_status_code} status code'
+        _custom_msg = _custom_msg.replace('failed', _status_code_msg)
+
+    # Construct the standard custom message if a custom message string was provided
+    if _message:
+        _custom_msg = f'{_custom_msg} {_message}'
+    elif _status_code:
+        # Adjust the status code message
+        _custom_msg = _custom_msg.split(_EXCEPTION_CLASSES._WITH_THE_FOLLOWING_SEGMENT)[0] + "."
+    else:
+        # Revert back to the default message if a custom message was not provided
+        _custom_msg = _EXCEPTION_CLASSES._API_DEFAULT_MSG.format(type=_request_type)
+
+    # Return the constructed message
+    return _custom_msg